refactored notificationScreen code and fixed bugs

+++ src/components/atoms/closeButton.js
@@ -1,26 +1,30 @@
-import {View, TouchableOpacity} from 'react-native';
-import React, {useEffect} from 'react';
+import {View, TouchableOpacity, Alert} from 'react-native';
+import React from 'react';
 import AntIcon from 'react-native-vector-icons/AntDesign';
 import {colorPalette} from './colorPalette';
-import {useDispatch, useSelector} from 'react-redux';
+import {useDispatch} from 'react-redux';
 import {loadDeleteNotification} from '../../redux/action/notification/deleteNotificationAction';
+import {loadGetAllNotification} from '../../redux/action/notification/getAllNotification';
 
-const CloseButton = ({notificationId, navigation}) => {
+const CloseButton = ({notificationId}) => {
   const dispatch = useDispatch();
-  const deleteNotificationdata = useSelector(
-    state => state.deleteNotificationReducer?.data,
-  );
-  console.log(deleteNotificationdata?.status, 'response');
 
   const deleteNotification = notificationId => {
-    dispatch(loadDeleteNotification(notificationId));
-    if (deleteNotificationdata?.status === 'Success') {
-      fetchAgain(navigation);
-    }
-  };
-
-  const fetchAgain = navigation => {
-    navigation?.navigate('Home');
+    Alert.alert('Are you sure?', 'Click ok to proceed', [
+      {
+        text: 'Ok',
+        onPress: () => {
+          dispatch(loadDeleteNotification(notificationId));
+          setTimeout(() => {
+            dispatch(loadGetAllNotification(0));
+          }, 2000);
+        },
+      },
+      {
+        text: 'Cancel',
+        onPress: () => {},
+      },
+    ]);
   };
 
   return (
